Add fix for broken pipe (#410)
grep --quiet, when used as part of a bash pipeline, closes the pipe and
sends the emitting process a SIGPIPE when it encounters the first
instance of a matching string. The aws process does not appropriately
handle a SIGPIPE and exits with an error, causing the script to fail.
This removes the -q in order to prevent grep from exiting early.