Specify namespace explicitly while creating ops-manager resource (#1176)
namespace is not mentioned in the docs which creating ops-manager resource which means if not explicitly specified
it will be created in the "default" namespaces. on the other hand, the docs mention specifying namespace for creating the admin-secret in the user-specified namespace. This change adds the command to README to create the ops-manager resource in a choice of namespace as well.