combine om init and om into one static image (#3497)
# Summary

as defined in the spec:
REDACTED

starting 7.0.4 (to be releasted) and 6.0.23 (need to be retagged or we
need to target the next patch version) we are combining om and om-init.

moving this change out of :
https://github.com/10gen/ops-manager-kubernetes/pull/3464/files#diff-c3d457979df06074195fa619e103a20f03b72f572231afd40a22e120b8f3b67a
which shows that it works

## Documentation changes

* [ ] Add an entry to [release notes](.../RELEASE_NOTES.md).
* [ ] When needed, make sure you create a new [DOCSP
ticket](https://jira.mongodb.org/projects/DOCSP) that documents your
change.

## Changes to CRDs

* [ ] Add `slaskawi`(Sebastian) and `@giohan` (George) as reviewers.
* [ ] Make sure any changes are reflected on `/public/samples`
directory.