CLOUDP-230125 - Make om static - remove init container (#3464)
# Summary

Prior PR to merge both images:
https://github.com/10gen/ops-manager-kubernetes/pull/3497

Spec and Discuss meeting related to this topic:
https://docs.google.com/document/d/1rsj_Ng3IRGv74y1yMTiztfc0LEpBVizjGWFQTaYBz60/edit#heading=h.6vtv3apus0s6

This PR leverages prior pr.

This pr does the following:
- on om image builds we add the init-container binaries
(backup_readiness, mms_configuration, docker-entry-point.sh,
backup-daemon-liveness-probe.sh) into the om container. That is because
changes are unlikely + they are operator independent and more likely to
be related to OM then to the operator
- we don't build a matrix by tying the operator version with om version
but instead we tie those binaries/scripts with each om version instead
(we can add a dedicated changelog if required, but changes are unlikely)

## How does it work for static and non static architecture?
- in static we don't use the init container and mounts, we directly
access the files under `/opt/scripts`
- in non static we do the same as before which means we mount on
`/opt/scripts` and copy files from the init-om-container to it. That
means everything which was there is over-mounted and over-written

-> All of that to say there should be no visible difference to the
customer as seen in the patches below

## Test
- full green patch:
https://spruce.mongodb.com/version/661564afdf7acf0007befa23/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

- static om without init container:
https://parsley.mongodb.com/taskFile/ops_manager_kubernetes_e2e_static_om70_kind_ubi_e2e_om_appdb_scram_patch_91367e01571aa4148535d4c281012219404f8961_6603ee61705d1d0007ac0730_24_03_27_10_01_07/0/om-scram-0-pod-describe.txt?bookmarks=0,96&shareLine=19
- non static om with init container:
https://parsley.mongodb.com/taskFile/ops_manager_kubernetes_e2e_om70_kind_ubi_e2e_om_appdb_scram_patch_91367e01571aa4148535d4c281012219404f8961_6603ee61705d1d0007ac0730_24_03_27_10_01_07/0/om-scram-0-pod-describe.txt?bookmarks=0,118&shareLine=19