CLOUDP-212055: Sign all images (#3505)
# Sign all images

## Introduction

This PR introduces signature for all container images released for the
enterprise operator, as well as daily rebuilds.

The signing process is standardised accros the whole org, with tools
provided by Devprod. The documentation is here:
https://docs.devprod.prod.corp.mongodb.com/release-tools-container-images/garasign/garasign_signing/

The Devprod team provides credentials for Garasign, a security
orchestration platform, so that we never handle our RSA keys directly.
This is done for each team.
Our public key is available at this address: https://cosign.mongodb.com/
, such that anyone can verify our binaries. They also provide a docker
container made for signing binaries and images.
We store the credentials for Garasign and the docker registry in our CI
secrets.

We decided not to implement that feature in Sonar, even if it would have
been a good fit for the tool, because we would like to get rid of it
gradually.

## How it works

After building and pushing an image or manifest to a registry, we fetch
its digest (this is not as triavial as we would expect and we use a tool
named `skopeo` for that).
Then we use a container image provided by DevProd to sign this image
digest with cosign, and the signature is pushed to the same registry.
We also do a signature verification step against our public key.

All images built during releases (`-context`) are signed, as well as
daily built images.

## Examples

There is now a new logo on Quay next to images that are signed. On ECR,
it is displayed as another file in the registry, next to the images,
named `<digest>.sig`.

Enterprise operator:

https://quay.io/repository/mongodb/mongodb-enterprise-operator-ubi?tab=tags
for example, tag `1.17.0-b20240411T124000Z`
Community operator:
https://quay.io/repository/mongodb/mongodb-kubernetes-operator?tab=tags
`0.8.3-b20240409T133900Z`
Init database:

https://quay.io/repository/mongodb/mongodb-enterprise-init-database-ubi?tab=tags
`1.0.16-b20240411T124000Z`
Init appdb on ECR:

https://us-east-1.console.aws.amazon.com/ecr/repositories/private/268558157000/images/ubi/mongodb-enterprise-init-appdb-ubi?region=us-east-1
`sha256-` files
Example of daily builds execution on Evergreen with signatures:

https://spruce.mongodb.com/version/66154520fa6ec700077c9f77/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

## Other changes

### Daily builds refactoring

- Lot of duplication in `daily.yaml` inventory removed, introduced
`architecture_suffix` parameter.
- Refactored the `build_image_daily` function, by creating
sub-functions.
- Added tests

Changes were originally part of a [separate
PR](https://github.com/10gen/ops-manager-kubernetes/pull/3468), but it
was a mess to rebase that one on top of it.

### Context images build refactoring
I noticed a lot of redudancy in the pipeline code while working on this
feature, nd adding signing further increased the amount of duplication,
so I refactored the functions for building context images (init appdb,
init database, operator...)

For example multiple functions contained these lines, and similar setup
for the `args` variable:
```
registry = QUAY_REGISTRY_URL + f"/mongodb-enterprise-{image_name}"
args["quay_registry"] = registry

sonar_build_image(image_name, build_configuration, args)
if build_configuration.sign and is_release_step_executed(build_configuration.get_skip_tags(), build_configuration.get_include_tags()):
    sign_image(registry, version + "-context")
```

This adds a lot to the scope of this PR, do not hesitate to ask me more
details if needed for reviewing.

### Logging
I improved the logger of the file, and replaced all `print()`statement
with calls to the logger.
The logger is configured to log `DEBUG` and `INFO` messages to `stdout`,
and above to `stderr`. Otherwise it logs everything to stderr and all
logs are displayed in red in most UIs.
I also changed the default log level to `DEBUG`, as they are most of the
time useful and the pipeline doesn't generate too many messages.

### QUAY_REGISTRY_URL variable
This new global var enables to override the base repository to which
images are pushed, otherwise they are hardcoded in inventory files and
it is hard to test changes without pushing to Quay.

## Follow up work

### Signing community images
Once I get feedback on this PR, I will implement the same features in
the community pipeline.

### Make bash execution from Python more readable
As discussed last Thursday, I will investigate if the usage of libraries
like [PyInvoke](https://www.pyinvoke.org/) can make the part with bash
script execution more readable.