CLOUDP-245961: Align containerImage in CSV annotations with deployment image (#3570)
Recently we've added a sanity check that is verifying if in CSV we have
both operator version consistent:
- `metadata.annotation.containerImage`
- deployment's image

Unfortunately containerImage is only set in Release PR by PCT and we
don't update it at all. For the release it is OK, but in tests, where we
override deployment's image to ECR one, we have inconsistency and fail.

With this patch we're making sure that
`metadata.annotation.containerImage` is always set to whatever is in
deployment.image when we prepare OLM bundles.