support creating multicluster config creator locally on darwin and ma… (#3608)
…ke e2e as well

# Summary
when running
`scripts/evergreen/build_multi_cluster_kubeconfig_creator.sh` (which
gets run as part of make prepare e2e locally and on evergreen as part of
the setup section) we create a binary of the multi-cluster tool based on
the current arch of your machine.

It will get copied to the test pod and used there (see the change in the
dockerfile). The problem is that the target archiecture relies on the
machine which it has been build for. Therefore, building it locally will
build it for darwin, copy it over and the test will fail.

I suggest instead that we build it 2 times:
- target architecture of your machine which gets cp to your kubectl
plugin directory
- linux architecture which will get copied to the test pod

in the worst case we are building it twice for the same architecture
(linux), but with this we are safe and the cost is neglible.

## Documentation changes

* [ ] Add an entry to [release notes](.../RELEASE_NOTES.md).
* [ ] When needed, make sure you create a new [DOCSP
ticket](https://jira.mongodb.org/projects/DOCSP) that documents your
change.

## Changes to CRDs

* [ ] Add `slaskawi`(Sebastian) and `@giohan` (George) as reviewers.
* [ ] Make sure any changes are reflected on `/public/samples`
directory.