Bump cluster cleaner on new Openshift cluster (#3636)
# Summary

This patch changes the default schedule for cleaning old testing
namespaces on openshift clusters to every 90 minutes for the upgrade
tests to have enough time to finish. The patch also ensures that OM
resources are deleted in the namespace as well. This PR is open to test
the updated credentials for the hosted openshift cluster which is
running on 4.15.

## Documentation changes

* [ ] Add an entry to [release notes](.../RELEASE_NOTES.md).
* [ ] When needed, make sure you create a new [DOCSP
ticket](https://jira.mongodb.org/projects/DOCSP) that documents your
change.

## Changes to CRDs

* [ ] Add `slaskawi`(Sebastian) and `@giohan` (George) as reviewers.
* [ ] Make sure any changes are reflected on `/public/samples`
directory.