CLOUDP-246692: Always add `=` to agent flags (#3642)
# Summary

This patch ensures that users can pass boolean flags to the startup
options of the agent in an intuitive fashion, e.g.:
```
spec:
  agent:
    startupOptions:
        someBooleanFlag: "false"
```
where before something like
```
spec:
  agent:
    startupOptions:
        someBooleanFlag: ""
```
was required to enable a flag and a flag that defaults to `true` was
only possible to override by doing:
```
spec:
  agent:
    startupOptions:
        someBooleanFlag=false: ""
```

This patch also handles the default value of
`-tlsRequireValidMMSServerCertificates` which defaults to true and
becomes impossible to override since we rely on environment variables to
the database pod to enable it.

## Documentation changes

* [ ] Add an entry to [release notes](.../RELEASE_NOTES.md).
* [ ] When needed, make sure you create a new [DOCSP
ticket](https://jira.mongodb.org/projects/DOCSP) that documents your
change.

## Changes to CRDs

* [ ] Add `slaskawi`(Sebastian) and `@giohan` (George) as reviewers.
* [ ] Make sure any changes are reflected on `/public/samples`
directory.