CLOUDP-252764 - fix flaky webhook tests by waiting for the hook to be ready first (#3695)
# Summary

Manually checking honeycomb it seems that all errors to webhook are most
likely due to the hook not being ready.
Example:
https://spruce.mongodb.com/task/ops_manager_kubernetes_e2e_static_multi_cluster_om_appdb_e2e_om_appdb_validation_patch_d3fe55d44a3aec9c3657518bb217a25c6df82bb6_669a507cd026ff0007e06c5a_24_07_19_11_39_45/files?execution=1&sortBy=STATUS&sortDir=ASC
shows in the `diagnostics.txt` that the hook worked as expected but the
test failed, which assumes some testing race

- Interestingly, `e2e_olm_operator_upgrade` is affected but since we
were not able to upload the logs we were not sure. Now we know that this
is the root cause...

## The flaky tasks

REDACTED
![Screenshot 2024-07-24 at 11 04
48](https://github.com/user-attachments/assets/2a615a5d-5cac-4c28-b553-b072772dfc56)

## Patch
Patch triggered with retries:
https://spruce.mongodb.com/version/66a0deaacc8e05000725f800/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

Another patch:
https://evergreen.mongodb.com/version/66a0f317427ede00076ab65f?redirect_spruce_users=true