Fix and improve check precommit (#3707)
# Summary

I noticed that recently we were pushing changes to master that were not
formatted with the pre commit hook, meaning that the `lint_repo` step of
the pipeline was broken, it was not failing when expected.
For example, on this patch (on master), Black reformats 4 files, but the
task succeeds:

https://spruce.mongodb.com/task/ops_manager_kubernetes_go_unit_tests_lint_repo_e209a7c60b973a1c2a9e94ea1b2ffe55784f94f8_24_07_30_12_53_39/logs?execution=0

This PR fixes this behaviour

Example of the task failing as expected, on an improperly formatted
code:

https://spruce.mongodb.com/task/ops_manager_kubernetes_go_unit_tests_lint_repo_patch_de551ea1f8b4431a6999564e5674572e69b5877b_66b2075c1c4cce0007e0942f_24_08_06_11_22_09/logs?execution=0

Execution after formatting is fixed (success):

https://spruce.mongodb.com/task/ops_manager_kubernetes_go_unit_tests_lint_repo_patch_de551ea1f8b4431a6999564e5674572e69b5877b_66b208d8062e7d0007082479_24_08_06_11_28_38/logs?execution=0

Note: Changes to files other than `check_precommit.sh` are only
formatting.