CLOUDP-249012: Fix project retrieval when project prefixes overlap (#3720)
# Summary

This patch fixes an error when the project look-up throws an error when
an in-exact project name match was found, as illustrated in the modified
E2E test:
- Previously, if trying to create `projectA` and `projectABC` already
exists, the project lookup would raise an error because the API would
return `projectABC` and when looping over the response, no exact match
for `projectA` would be found and instead of returning an empty project
to be created, the operator would raise an error and `projectA` would
fail to be created.

## Documentation changes

* [x] Add an entry to [release notes](.../RELEASE_NOTES.md).
* [ ] When needed, make sure you create a new [DOCSP
ticket](https://jira.mongodb.org/projects/DOCSP) that documents your
change.

## Changes to CRDs

* [ ] Add `slaskawi`(Sebastian) and `@giohan` (George) as reviewers.
* [ ] Make sure any changes are reflected on `/public/samples`
directory.