Add finalizer to the MogoDBUser resource (#3715)
# Summary

A `finalizer` is a string that is added to the list in the
`metadata.finalizers` field of the object. When the object is deleted,
Kubernetes checks for the `metadata.finalizers` field and if it is not
empty, it does not proceed with the deletion. Instead it marks the
object for deletion by adding a new field `metadata.deletionTimestamp`.
After this point the object becomes read only except for the
`metadata.finalizers` field. When the field is emptied of all
finalizers, Kubernetes resumes the deletion.

This PR adds finalizers to the MongoDBUser resource. This will make sure
that the resource is only deleted after the user has been removed from
the `AutomationConfig` such that it is not possible for users to remain
active after the resource is removed from the k8s cluster.

To implement finalizers the following approach was used. When the
MongoDBUser resource is created, in the `Reconcile` loop, the finalizer
will be added to the resource if it is not present yet. When the
resource is deleted(which is checked by the presence of
`deletionTimestamp`, in the `Reconcile` loop), the pre deletion cleanup
will be performed(in this case the update of the `AutomationConfig`).
After this is completed successfully, the finalizer is removed and the
`delete` method is triggered automatically.

The PR:
- Adds logic for adding and removing the finalizer in the
`controllers/mongodbuser_controller.go`
- Adds 2 unit tests in `controllers/mongodbuser_controller_test.go` to
test the implemented functionality
- `TestFinalizerIsAdded_WhenUserIsCreated` checks that the finalizer has
been added when the user is created
- `TestFinalizerIsRemoved_WhenUserIsDeleted` checks that the finalizer
has been removed when the user is deleted
- Adds 2 new e2e tests in `docker/mongodb-enterprise-tests/tests/users`
- `users_finalizer.py` that checks the expected behaviour of finalizers
and clean up logic
- `users_finalizer_removal.py` that checks that the user can be deleted
after the referenced MongoDB resource has been deleted
- Adds a new fixture for the e2e test in
`docker/mongodb-enterprise-tests/tests/users/fixtures`,
`user_scram.yaml`

## Documentation changes

* [ ] Add an entry to [release notes](.../RELEASE_NOTES.md).
* [ ] When needed, make sure you create a new [DOCSP
ticket](https://jira.mongodb.org/projects/DOCSP) that documents your
change.

## Changes to CRDs

* [ ] Add `slaskawi`(Sebastian) and `@giohan` (George) as reviewers.
* [ ] Make sure any changes are reflected on `/public/samples`
directory.