Fixed failing OLM tests at release time (#3734)
We had a mismatch between prepare OLM bundles step and the actual OLM
upgrade tests in how the latest released version was obtained and what
was the current version of the operator.
At release time, when release.json had bumped versions, but not released
yet, there were ImagePullBackoff errors.

Now, both prepare and test calculate the versions the same way:

- latest released version is taken from the certified operators repo
- current version is taken from release.json