Use namespace as prefix for project names in e2e test (#3750)
# Summary

This fixes the `replicaset_agent_flags` test creating organizations with
a name pattern that is not recognized by the periodic teardown to allow
us to not keep creating those projects. A proper fix that will ensure
that the 2 orgs created in the test are removed at the end of the test
will follow.

EVG run:
https://spruce.mongodb.com/version/66d5b34eae5d2f0007884e3b/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

## Documentation changes

* [ ] Add an entry to [release notes](.../RELEASE_NOTES.md).
* [ ] When needed, make sure you create a new [DOCSP
ticket](https://jira.mongodb.org/projects/DOCSP) that documents your
change.

## Changes to CRDs

* [ ] Add `slaskawi`(Sebastian) and `@giohan` (George) as reviewers.
* [ ] Make sure any changes are reflected on `/public/samples`
directory.