CLOUDP-203244 - Set default security context for init containers (#3760)
# Summary

This PR fixes the bug detailed in
[CLOUDP-203244](https://jira.mongodb.org/browse/CLOUDP-203244) and
[HELP-50718](https://jira.mongodb.org/browse/HELP-50718). There is a
default security context which is set for every pod, and for every
container. They are set at 2 different levels since the pod-level
security context can't set `allowPrivilegeEscalation` and
`readOnlyRootFilesystem`, so they need to be set at container level.
As the ticket mentions, the default security context is not set properly
at container level for the init containers, leading to azure policy
violations. This affected appdb and mongodb deployments (only in
non-static architecture).

Added unit tests to check that the security context field is set
properly. Did not add any e2e test since this was affecting the yaml
applied to the cluster and not the behavior afterwards, but let me know
if I should add any.

## Documentation changes

* [ ] Add an entry to [release notes](.../RELEASE_NOTES.md).
* [ ] When needed, make sure you create a new [DOCSP
ticket](https://jira.mongodb.org/projects/DOCSP) that documents your
change.

## Changes to CRDs

* [ ] Add `slaskawi`(Sebastian) and `@giohan` (George) as reviewers.
* [ ] Make sure any changes are reflected on `/public/samples`
directory.