Switch context with fzf (#3837)
This is a very simple helper to switch contexts more easily. 

Often we want to switch to some context but don't know the exact name
and we don't have the switch command in the history. So we go to IDE,
find a context file, rename the file, copy the name, go to terminal and
paste the name as context= parameter (at least that was my way of coping
with that 🙄 ).

Now, you can just write `make context` and you'll get fzf-powered
context picker. You want context for om multi-cluster with static
containers? Write "om multi static" and it will fuzzy filter and narrow
down considerable what you probably want.

This will only trigger fzf if no context is passed. If you specify
context, everything works as before.


![switch](https://github.com/user-attachments/assets/a744a8b2-68d3-4182-99ac-debccb79f1fa)