CLOUDP-275846 Fix the diagnostic support script (#3848)
# Summary

The diagnostic support script doesn't handle the case where both the
MongoDB and MongoDBOpsManager resources are deployed in the same
namespace because of the way it enumerates pods it thinks belong to a
MongoDB deployment. This adds a second filtering stage to weed out
non-MongoDB pods.

This could have been implemented with a jsonpath filter on the first
`kubectl get pods` invocation but the limitation there is that the
filter could only be expressed in terms of a specific container's name
at the n-th position, not any position, which can break if container
ordering changes for any reason.

## Documentation changes

* [ ] Add an entry to [release notes](.../RELEASE_NOTES.md).
* [ ] When needed, make sure you create a new [DOCSP
ticket](https://jira.mongodb.org/projects/DOCSP) that documents your
change.

## Changes to CRDs

* [ ] Add `slaskawi`(Sebastian) and `@giohan` (George) as reviewers.
* [ ] Make sure any changes are reflected on `/public/samples`
directory.