CLOUDP-280290: Refactor config servers and mongos configurations in controller (#3864)
# Summary

This pull requests refactor the logic for handling Mongos and Config
servers configuration and overrides.
We create methods similar to `prepareDesiredShardsConfiguration`, to
merge everything in one place, and end up with a unified data structure
of type `ShardedClusterComponentSpec`.

Main changes:

- Added two new functions `prepareDesiredMongosConfiguration` and
`prepareDesiredConfigServerConfiguration`. using as much as shared logic
as possible. Mainly through function `processClusterSpecList`, that was
already used for `prepareDesiredShardsConfiguration`.
- "get options" functions for mongos and config now take the desired
configuration as a parameter. e.g `ConfigServerOptions`. The logic of
these functions has been refactored in `shardedCommonOptions`.
- Adapt existing unit tests to work with the new methods signatures.
- Create two unit tests for the new prepareDesired functions. The
previous method used for testing expected shard configuration was
re-used and refactored in `testDesiredConfigurationFromYAML`
- The mongodb resource and associated expected configuration are stored
as YAML file. e.g `mdb-sharded-multi-cluster-configsrv-mongos.yaml` ;
mongos and config servers share the same logic, so they use the same
files.

Config servers and Mongos configuration fields are the following:

![image](https://github.com/user-attachments/assets/f5352d92-4f94-4138-a265-45bd417c0504)