CLOUDP-207592 Rotate TLS Certificates Safely (#3852)
# Summary

This pull request improves the way we do certificate rotation by keeping
the previous certificate available in the pods. This fixes the deadlock
that can appear between the agents and the automation config (the agents
are not ready because they don't have the latest certificate from the
automation config, and the automation config is not updated because the
agents are not ready).

Changes:
* Main change in in `CreatePemSecretClient`. This method now has a flag
which enables or disables the functionality of "safeRotation". This flag
exists because certificates for the agent don't benefit from this since
they are not stored under a certificate hash, but a hardcoded key.
* The "safeRotation" functionality resides in
`updateSecretDataWithLatestHash`. This method compares the new
certificate received with the data already present in the `-pem` secret.
Using a new key `latestHash` we can determine which certificates to put
in the final `-pem` secret. Also added a `previousHash` key in the
secret to use for vault.
* Created e2e tests which simulate this deadlock by triggering a
statefulset restart. Tests have been created for every component:
replicasets, sharded clusters (mongos, configsrv, shard), ops manager,
appdb.
* Updated some e2e test that were verifying certificate rotation. These
tests were instantly passing because the mdb resource stays in a
`Running` phase for a while after a rotation happens. Added a
`assert_abandons_phase` to make sure that the rotation happens
successfully.
* Removed leftover code and fixed tests from the time when tls
certificates were allowed to be generic. TLS certificate have to be of
the `kubernetes.io/tls` type (have been for a while), but there were
some leftovers and some unit tests.
* Updated hashicorp vault annotations to not mount all the values in a
secret, but only the key we require.
* Added new hashicorp vault annotations to mount the previous
certificate when using this as the secret backend.

## Documentation changes

* [x] Add an entry to [release notes](.../RELEASE_NOTES.md).
* [ ] When needed, make sure you create a new [DOCSP
ticket](https://jira.mongodb.org/projects/DOCSP) that documents your
change.

## Changes to CRDs

* [ ] Add `slaskawi`(Sebastian) and `@giohan` (George) as reviewers.
* [ ] Make sure any changes are reflected on `/public/samples`
directory.