Fix daily build not building all images (#3917)
# Summary

Due to a bug introduced in #3883 Sonar was only actually building the
last image in the matrix, over and over. The fix is to always copy the
`args` dict that contains the requested version image before submitting
it to a concurrent Executor for pickling so that the next loop iteration
doesn't update it in place.

## Documentation changes

* [ ] Add an entry to [release notes](.../RELEASE_NOTES.md).
* [ ] When needed, make sure you create a new [DOCSP
ticket](https://jira.mongodb.org/projects/DOCSP) that documents your
change.

## Changes to CRDs

* [ ] Add `slaskawi`(Sebastian) and `@giohan` (George) as reviewers.
* [ ] Make sure any changes are reflected on `/public/samples`
directory.