CLOUDP-279249: Sharded disaster recovery test + Fix replicaset ordering in AC (#3934)
# Summary

This PR introduces an E2E test for a simple disaster recovery scenario:
we lose one cluster without losing the majority.
We ensure that the operator correctly ignores the unhealthy cluster in
the subsequent reconciliation, and we can still scale.

While writing this test, I discovered a bug in the way we update the
automation configuration. We hadn't kept track of the _id fields, so
when we rescaled, we changed a host's unique _id, and the agent doesn't
support that.

---------

Co-authored-by: Łukasz Sierant <lukasz.sierant@mongodb.com>