MC Sharded scaling unit tests (#3946)
Changes and fixes to Multi-Cluster Sharded scaling:
- fixed preparing shard hostnames
- fixed checking whether the multi-cluster replicaset (governed by a
list of scalers for each member cluster) is still undergoing scaling.
The problem was in DesiredReplicas method of the scaler, which
incorrectly reported desired scaling if more than one member cluster was
scaled at the same time. DesiredReplicas works correctly to limit
scaling to only one member of a replicaset (across multiple statefulsets
in multiple member clusters), but it cannot be used for checking if any
other statefulsets in other clusters are still to be scaled. For this
case (reporting the correct desired replicas in logs and for checking if
there is anything still to be scaled) TargetReplicas method was added to
the scaler.
- improved logging the scaling status
- improved reporting which agent hostnames the operator is waiting for
(printing expected, actual and missing hostnames in logs)
- propagating the detailed hostname message (described above) to the CR
status