Sign and doing sbom in parallel (#3963)
# Summary

This works tries to run per set of agent images the following steps
concurrently:
- creating sbom
- signing images
But all of them are done **after** the image has been created.
- removal not used agent
- adding retry for warning, as sometimes after pushing the signature,
the signature is still not available.

Signing the image takes around 3-5 minutes. Which is a significant time
investment.

- Example speed up ->
[1h](https://spruce.mongodb.com/version/67505bb4794c34000700fa73/tasks?page=0&sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)
vs
[2h](https://spruce.mongodb.com/version/675048333b14230007acd928/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)

the different to
[prior](https://github.com/10gen/ops-manager-kubernetes/pull/3883/files)
parallelisation effort is that we only run sbom and signing concurrently
- while the image creation stays sequentially.

periodic rebuild for normal agent worked:
https://spruce.mongodb.com/version/67517162fdc4670007882ebb/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC