flaky: fix e2e_multi_cluster_recover by moving transition check (#4003)
# Summary

- `e2e_multi_cluster_recover` is most of the time flaky. It fails on the
assumption that the resource will go through a state change if the
operator has been started. That is racy by nature as we removed a lot of
state transition code. A better solution would be to save the
transitionDate before restarting the operator. Having said that, the
generation check should take that into consideration already

https://spruce.mongodb.com/task/ops_manager_kubernetes_e2e_multi_cluster_kind_e2e_multi_cluster_recover_network_partition_patch_f16faa0aa0926cc89f02a25211643ee1b123e85c_6781170e6317f00007bde1bc_25_01_10_12_48_15/logs?execution=0

![Screenshot 2025-01-10 at 14 33
07](https://github.com/user-attachments/assets/7cf8f5c6-e92c-4722-b72c-4b981ca5db70)