flaky: add retries to cert validation and ignore intermediate failed phase (#4019)
# Summary

The flaky
[test](https://spruce.mongodb.com/task/ops_manager_kubernetes_e2e_static_operator_kind_ubi_cloudqa_e2e_operator_upgrade_sharded_cluster_patch_13354d1c1bf32d38368a5f19ceef1096eef3b5aa_6787727366d83c00070f38ad_25_01_15_08_31_52/logs?execution=0&page=0&sortBy=STATUS&sortDir=ASC)
first deploys the 1.27.0 operator, then upgrades to current latest, and
then downgrades back to 1.27.0. The epic on TLS improvements was
released in 1.29.0, which added new keys to the pem secrets. Because of
that, during the downgrade, the 1.27 operator doesn't recognize those
keys and gives this error
```
The Secret prefix-sh001-base-0-cert-pem containing certificates is not valid. Entries must contain a certificate and a private key.
```
This moves the resource to Failed briefly before continuing to reconcile
properly, but this fails the test. As a fix, we ignore the intermediate
Failed phases.

Additionally, the operator creates the pem secret and immediately after
that verify it. This can be racy. As a quick fix I've added retries.

## Proof
- green evergreen patch with 0 retries:
https://spruce.mongodb.com/task/ops_manager_kubernetes_e2e_static_operator_kind_ubi_cloudqa_e2e_operator_upgrade_sharded_cluster_patch_13354d1c1bf32d38368a5f19ceef1096eef3b5aa_678a685bd7fe6e0007638cd0_25_01_17_14_25_32/logs?execution=0

---------

Co-authored-by: Lucian Tosa <lucian.tosa@mongodb.com>
Co-authored-by: Lucian Tosa <49226451+lucian-tosa@users.noreply.github.com>