preflight: fix exception thrown for 'mongodb-enterprise-server' images (#4052)
# Summary

The following exception was thrown in the preflight logs:
```
ERROR:root:Preflight check failed with exception: 'set' object is not subscriptable
```
The preflight parallel method was trying to access a specific index in
the `available_versions` list. This is normally a list, but when the
image is `mongodb-enterprise-server` this is then a set therefore
throwing the exception.

This issue only affected preflight checks for the
`mongodb-enterprise-server` images. Preflight would still run, but the
task would not fail in case preflight failed because the exception was
thrown before collecting return codes.

## Documentation changes

* [ ] Add an entry to [release notes](.../RELEASE_NOTES.md).
* [ ] When needed, make sure you create a new [DOCSP
ticket](https://jira.mongodb.org/projects/DOCSP) that documents your
change.

## Changes to CRDs

* [ ] Add `slaskawi`(Sebastian) and `@giohan` (George) as reviewers.
* [ ] Make sure any changes are reflected on `/public/samples`
directory.