CLOUDP-294523 CLOUDP-294516 Pin OM version for multi-cluster DR test-cases (#4056)
# Summary

This patch pins the OM version we use in the multicluster DR tests to
`7.0.12` as this is the latest OM 7.0.x version that doesn't
automatically run pre-migration checks before every restart. Aligning
with EA on a path forward will lead to removing this pin.

Created https://jira.mongodb.org/browse/CLOUDP-297377 to track the
effort on our end to fix the pre-flight startup.

## Documentation changes

* [ ] Add an entry to [release notes](.../RELEASE_NOTES.md).

## Changes to CRDs

* [ ] Make sure any changes are reflected on `/public/samples`
directory.