CLOUDP-295444 - add proxy support for the operator (#4044)
# Summary

This patchs adds the following:
- [x] proxy env var on the container the operator deploys can be
propagated to the agent.
This is sufficient because both the operator and the agent take the
proxy environment variables into account.
- [x] Propagating the proxy environment variables is protected by the
`PROPAGATE_PROXY_ENV` environment variable.
This is necessary because it's valid to not propagate proxy environment
variables in a multi-cluster context by default.
- [x] e2e test with squid-proxy.
I used squid-proxy because of the availability of an image published by
a certified publisher. Nginx or writing a proxy from scratch is not
ideal for a generic forward proxy.

## Proof of Work

Successful EVG run:
https://spruce.mongodb.com/version/67a27126d70b050007296e75

## Checklist
- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you checked for release_note changes?

---------

Co-authored-by: Mircea Cosbuc <mircea.cosbuc@mongodb.com>