CLOUDP-300255 - Fix concurrency in SBOM generation (#4094)
# Summary

During SBOM generation we were spawning multiple threads which were
accessing the same `args` dictionary. At the same time we were modifying
`args` so all threads were getting this modification. This meant that
most times we were generating the SBOM for the same architecture twice,
and missing one of the architectures.

Why did this work until now? Building and generating SBOMs daily meant
that we will generate this for almost every image given enough time.

Additionally, also set the SSDLC script to generate reports for the
agent_operator images, only with the latest operator.

## Proof of Work
If you take a look in our SBOM bucket
https://us-east-1.console.aws.amazon.com/s3/buckets/kubernetes-operators-sboms?prefix=sboms/daily/augmented/quay.io/mongodb/mongodb-agent-ubi/107.0.13.8702-1/&region=us-east-1&bucketType=general

you can see that we were missing linux-arm64 SBOM on 11th of February,
but this SBOM was built twice

https://us-east-1.console.aws.amazon.com/s3/object/kubernetes-operators-sboms?region=us-east-1&bucketType=general&prefix=sboms%2Fdaily%2Faugmented%2Fquay.io%2Fmongodb%2Fmongodb-agent-ubi%2F107.0.13.8702-1%2Flinux-arm64%2Fsha256%3Ad67a4a67334fbb29c2afa8a361defd48b621ae3887d2397735c0e82e2bf86b3b&tab=versions

Also, checking the logs for the periodic build this morning and
searching for "generating sbom for"
https://parsley.mongodb.com/evergreen/ops_manager_kubernetes_periodic_build_periodic_build_agent_67ac1cc7bbbb000007a6fecf_25_02_12_04_00_07/0/task?bookmarks=0,15171

you will see that we generate an SBOM for the same image and
architecture twice. Doing the same for these logs (containing these
changes), this doesn't happen anymore

https://parsley.mongodb.com/evergreen/ops_manager_kubernetes_periodic_build_periodic_build_agent_patch_fb61b679b17819cdfca0479d3b5ef456d732f8c0_67ac7ce86f31f70007e4e4f4_25_02_12_10_50_17/0/task?bookmarks=0,3226

## Checklist
- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you checked for release_note changes?

## Reminder (Please remove this when merging)
- Please try to Approve or Reject Changes the PR, keep PRs in review as
short as possible
- Our Short Guide for PRs:
[Link](https://docs.google.com/document/d/1T93KUtdvONq43vfTfUt8l92uo4e4SEEvFbIEKOxGr44/edit?tab=t.0)
- Remember the following Communication Standards - use comment prefixes
for clarity:
  * **blocking**: Must be addressed before approval.
  * **follow-up**: Can be addressed in a later PR or ticket.
  * **q**: Clarifying question.
  * **nit**: Non-blocking suggestions.
  * **note**: Side-note, non-actionable. Example: Praise
  * --> no prefix is considered a question