CLOUDP-301110 Ensure E2E tests correctly run in static mode (#4111)
# Summary

The following tests were running in non-static mode despite having a
static context activated:
- `operator_upgrade_sharded_cluster`
- `multi_cluster_appdb_state_operator_upgrade_downgrade`
- `e2e_olm_operator_upgrade`
- `e2e_olm_operator_upgrade_with_resources`

To fix OLM tests, the variable `MDB_DEFAULT_ARCHITECTURE` is explicitly
passed to the operator env in the test.
OLM tests context was also updated to run with OM70.

For the two other tests, the variable is passed in the
`install_official_operator` method, as an helm argument.

## Proof of Work

In Evergreen, we see in the pod descriptions of deployed MongoDB
resources that the environment variables contain
`MDB_STATIC_CONTAINERS_ARCHITECTURE: true`, and agent images pulled are
suffixed with the operator version. In static variants only.

Example: [Sharded pod
description](REDACTED)